Add separate make targets for upstream examples (#1328)
* If the upstream examples are changed while working on a feature, it
can be frustrating to regenerate YAML files as they cannot match without
manually editing the Python versions, which mixes up intentions in the
PR.
* For that reason, use separate make targets for regenerating the
Hera-generated YAML files (`regenerate-upstream-test-data`) and fetching
the actual upstream YAML files - `fetch-upstream-examples`.
* Add a check (when `REPORT_DIFFS` env var is set) to ensure that the
local copy of the upstream YAML matches for
`regenerate-upstream-test-data`, with a test fail message to tell user
to fetch the new YAML with the `fetch-upstream-examples` target.
* Update the `regenerate-test-data` target to only regenerate
non-upstream examples.

With this change the new developer workflow will be:
* If making changes to Hera-specific examples (e.g. runner examples),
use `make regenerate-test-data`
* If making changes that affect all examples (e.g. YAML key ordering
#1326) which should rarely happen, use both: `make regenerate-test-data
regenerate-upstream-test-data`
* See if there are changes to upstream example YAMLs that have been
recreated in Hera by running `REPORT_DIFFS make
regenerate-upstream-test-data` (TODO: add an automated way to notify us
of changes)
* If there are new upstream examples or they have changed, run `make
fetch-upstream-examples` to update the `.upstream.yaml` files, then work
on the respective Python files to update the Hera workflow, and then
regenerate the matching `.yaml` files using `make
regenerate-upstream-test-data`

---------

+++ tests/test_examples.py
@@ -25,6 +25,7 @@ from hera.workflows.script import InlineScriptConstructor
 
 ARGO_EXAMPLES_URL = "https://raw.githubusercontent.com/argoproj/argo-workflows/main/examples"
 HERA_REGENERATE = os.environ.get("HERA_REGENERATE")
+REPORT_DIFFS = os.environ.get("REPORT_DIFFS")
 CI_MODE = os.environ.get("CI")
 
 LOWEST_SUPPORTED_PY_VERSION = (3, 9)
@@ -152,13 +153,29 @@ def test_hera_output_upstream(module_name, global_config_fixture):
     # WHEN
     output = workflow.to_dict()
 
-    # THEN - generate the yaml if HERA_REGENERATE or it does not exist
+    # THEN - generate the workflow yaml if HERA_REGENERATE or it does not exist
     if _generate_yaml(generated_yaml_path):
         generated_yaml_path.write_text(yaml.dump(output, sort_keys=False, default_flow_style=False))
+
     if _generate_yaml(upstream_yaml_path):
-        upstream_yaml_path.write_text(
-            requests.get(f"{ARGO_EXAMPLES_URL}/{module_name.replace('__', '/').replace('_', '-')}.yaml").text
-        )
+        if not upstream_yaml_path.exists():
+            # Only fetch and write upstream YAML if it doesn't exist (use
+            # `make fetch-upstream-examples` to update all existing examples)
+            upstream_yaml_path.write_text(
+                requests.get(f"{ARGO_EXAMPLES_URL}/{module_name.replace('__', '/').replace('_', '-')}.yaml").text
+            )
+        elif REPORT_DIFFS:
+            # Compare current local YAML with upstream from GitHub
+            remote_upstream_yaml = requests.get(
+                f"{ARGO_EXAMPLES_URL}/{module_name.replace('__', '/').replace('_', '-')}.yaml"
+            ).text
+            local_yaml = upstream_yaml_path.read_text()
+            if local_yaml != remote_upstream_yaml:
+                assert False, (
+                    f"Local copy of upstream YAML does not match actual upstream YAML for {module_name}"
+                    " fetch new YAML by running `make fetch-upstream-examples` (you may then need to edit"
+                    " the code in the Python file)."
+                )
 
     # Check there have been no regressions from the generated yaml
     assert generated_yaml_path.exists()
